pyproject.toml
[tool.poetry]
name = "upbit-stream-broadcaster"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.10.*"
pytz = "^2024.2"

[tool.poetry.group.dev.dependencies]
pyzmq = "^26.2.0"
pytest = "^8.3.3"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
